Separate HtmlSubgraph
