+++ backend/src/airas/features/publication/generate_html_subgraph/input_data.py
@@ -1,0 +1,1 @@
+generate_html_subgraph_input_data: dict = {}
