--- backend/src/airas/features/publication/html_subgraph/input_data.py
@@ -1,1 +1,0 @@
-html_subgraph_input_data = {}
